Move correlation plot to own component
